Internal: Code refactoring.
