+++ src/ckeditorinspector.css
